+++ R/apply_vaccination.R
@@ -136,6 +136,9 @@ apply_vaccination <- function(ngm,
     dplyr::pull(transmission_reduction_matrix)
 
   ngm_vaccinated <- Map("*", ngm, transmission_reduction_matrix)
-  ngm_vaccinated <- new_setting_vaccination_matrix(ngm_vaccinated)
+  ngm_vaccinated <- new_setting_vaccination_matrix(
+    ngm_vaccinated,
+    age_breaks = age_breaks(ngm)
+  )
   return(ngm_vaccinated)
 }