+++ R/constructors.R
@@ -80,21 +80,45 @@ age_breaks.conmat_setting_prediction_matrix <- function(matrix) {
 
 #' @describeIn age_breaks Get age break information
 #' @export
+age_breaks.setting_data <- function(matrix) {
+  attr(matrix, "age_breaks")
+}
+
+#' @describeIn age_breaks Get age break information
+#' @export
+age_breaks.ngm_setting_matrix <- function(matrix) {
+  attr(matrix, "age_breaks")
+}
+
+#' @describeIn age_breaks Get age break information
+#' @export
+age_breaks.setting_vaccination_matrix <- function(matrix) {
+  attr(matrix, "age_breaks")
+}
+
+#' @describeIn age_breaks Get age break information
+#' @export
 age_breaks.default <- function(matrix) {
-  cli::cli_abort("no method for {.code age_breaks()} defined yet")
+  cli::cli_abort("no method for {.code age_breaks()} defined for object of class {.cls {class(matrix)}} yet")
 }
 
 new_setting_data <- function(list_df) {
-  add_new_class(list_df, "setting_data")
+  structure(
+    list_df,
+    age_breaks = unique(list_df$home$age_from),
+    class = c("setting_data", class(list_df))
+  )
 }
 
 new_ngm_setting_matrix <- function(list_matrix,
                                    raw_eigenvalue,
-                                   scaling) {
+                                   scaling,
+                                   age_breaks) {
   structure(
     list_matrix,
     raw_eigenvalue = raw_eigenvalue,
     scaling = scaling,
+    age_breaks = age_breaks,
     class = c("ngm_setting_matrix", class(list_matrix))
   )
 }
@@ -152,6 +176,11 @@ new_setting_contact_model <- function(list_model) {
   add_new_class(list_model, "setting_contact_model")
 }
 
-new_setting_vaccination_matrix <- function(list_matrix) {
-  add_new_class(list_matrix, "setting_vaccination_matrix")
+new_setting_vaccination_matrix <- function(list_matrix,
+                                           age_breaks) {
+  structure(
+    list_matrix,
+    age_breaks = age_breaks,
+    class = c("setting_vaccination_matrix", class(list_matrix))
+  )
 }